requirements.txt
# 节点运行所需依赖
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.24.0
soundfile>=0.12.1
torch>=2.0.0